Move upgrade repository test function
